pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.0",
    "pytest-cov>=7.1.0",
    "pyyaml>=6.0.3",
    "ruff>=0.15.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[tool.black]
line-length = 80
target-version = ["py312"]
skip-string-normalization = false

[project.scripts]
gendiff = "gendiff.main:main"
